pyproject.toml
[project]
name = "cloud-workflows-generator"
version = "0.1.0"
description = "Pydantic v2 YAML validator for Google Cloud Workflows syntax"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
cloud-workflows = "cloud_workflows.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.pytest.ini_options]
testpaths = ["tests"]
